[Fixes #106731208] refactor: renamed schema folder to models and corrected referencing of the models in the app's business logic
