+++ server/controllers/items.js
@@ -1,7 +1,7 @@
 (function() {
   'use strict';
 
-  var Items = require('../schemas/items');
+  var Items = require('../models/items');
 
   module.exports = {
     // Middleware to create an item
